+++ CHANGELOG.md
@@ -9,8 +9,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ## [Current nightly]
 
 ### Added
+- Added `-Experimental` to `Get-PnPSiteTemplate`, `Invoke-PnPSiteTemplate`, `Get-PnPTenantTemplate`, `Invoke-PnPTenantTemplate`, `Read-PnPSiteTemplate`, `Save-PnPSiteTemplate`, `Read-PnPTenantTemplate`, `Save-PnPTenantTemplate`, `Convert-PnPSiteTemplate`, `Export-PnPListToSiteTemplate`, `Add-PnPDataRowsToSiteTemplate`, `Add-PnPFileToSiteTemplate`, `Remove-PnPFileFromSiteTemplate` and `Set-PnPSiteTemplateMetadata` to run them on the experimental PnP.Core.Provisioning engine instead of PnP Framework. 
 
 ### Changed
+- Changed `-InputInstance` on `Invoke-PnPSiteTemplate` and `-Template` on `Invoke-PnPTenantTemplate` to accept a template produced by either provisioning engine, so a template extracted with `-Experimental` can be applied with `-Experimental`. Templates produced by PnP Framework keep binding to these parameters as before.
 - Telemetry in PnP PowerShell has been removed due to the costs of collecting the data didn't outweigh the benefits to the PnP PowerShell team to have insights into its usage. The involved cmdlets `Get-PnPPowerShellTelemetryEnabled`, `Enable-PnPPowerShellTelemetry` and `Disable-PnPPowerShellTelemetry` have been marked as deprecated and no longer function, but will stay in v3 for backwards compatibility with existing scripts. These cmdlets will be removed in the next v4 release. All versions of PnP PowerShell will no longer be able to submit telemetry. You might see background requests for this failing. This will not interfear with the normal execution of your PowerShell script. [#5460](https://github.com/pnp/powershell/pull/5460)
 
 ### Fixed